+++ src/components/NotFound/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+NotFound.propTypes = {};
+
+function NotFound(props) {
+    return <div>Not Found</div>;
+}
+
+export default NotFound;
