+++ src/features/Todo/pages/DetailPage/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+DetailPage.propTypes = {};
+
+function DetailPage(props) {
+    return <div>Detail Page</div>;
+}
+
+export default DetailPage;
